created controller method to return products by category

+++ controllers/product.controller.js
@@ -24,6 +24,9 @@ class ProductController {
     static getAllProducts = async(req, res, next)=>{
         try{
             let product = await Product.find();
+            if(req.query.categoryID){
+                return getProductsByCategory(req, res, req.query.categoryID)
+            }else if(Object.keys(req.query).length > 0) throw new Error("Not a valid query parameters");
             JSONResponse.success(res, "Successfully created product", product, 200);
         }catch(error){
             JSONResponse.error(res, "Unable to retrieve product", error, 404);
@@ -52,6 +55,8 @@ class ProductController {
         try{
             let id = req.params.id;
             let product = await Product.findById(id);
+            if(!product) throw new Error("Product not found with that ID")
+
             if(product.prod_img){
                 await awsStorage.deleteObjectFromS3(product.prod_img);
              };
@@ -66,9 +71,22 @@ class ProductController {
         try{
             let id = req.params.id;
             let product = await Product.findById(id);
+            if(!product) throw new Error("Product not found with that ID")
             JSONResponse.success(res, "Product successfully retrieved", product, 200);
         }catch(error){
             JSONResponse.success(res, "Failed to retrieve product", error, 404)
+        }
+    }
+
+    static getProductsByCategory = async(req, res, categoryID)=>{
+        try{
+            if(categoryID){
+                let products = await Product.find({category: categoryID});
+                JSONResponse.success(res, "Products successfully retrieved", products, 200);
+            }
+            
+        }catch(error){
+            JSONResponse.success(res, "Failed to retrieve products", error, 404)
         }
     }
